+++ packages/browser-destinations/src/destinations/heap/generated-types.ts
@@ -22,7 +22,7 @@ export interface Settings {
    */
   hostname?: string
   /**
-   * This is an optional setting. When set, nested arrays items will be sent in as new Heap events. Defaults to 0.
+   * This is an optional setting. When set, nested array items will be sent in as new Heap events. Defaults to 0.
    */
   browserArrayLimit?: number
 }
